add loan application summary chart

+++ API/Loan/Loan.Framework/Reporting/LoanApplicationSummaryItem.cs
@@ -2,7 +2,7 @@
 {
     public readonly struct LoanApplicationSummaryItem
     {
-        public LoanApplicationSummaryItem(short year, short month, int count, string description)
+        public LoanApplicationSummaryItem(int year, int month, int count, string description)
         {
             Year = year;
             Month = month;
@@ -10,8 +10,8 @@
             Description = description;
         }
 
-        public short Year { get; init; }
-        public short Month { get; init; }
+        public int Year { get; init; }
+        public int Month { get; init; }
         public int Count { get; init; }
         public string Description { get; init; }
     }
