+++ API/Loan/Loan.Reporting.Data/Models/LoanApplicationCloseData.cs
@@ -2,8 +2,8 @@
 {
     public class LoanApplicationCloseData
     {
-        [ColumnMapping] public short ClosedYear { get; set; }
-        [ColumnMapping] public short ClosedMonth { get; set; }
+        [ColumnMapping] public int ClosedYear { get; set; }
+        [ColumnMapping] public int ClosedMonth { get; set; }
         [ColumnMapping] public string StatusDescription { get; set; }
         [ColumnMapping] public int Count { get; set; }
     }
