+++ API/Loan/Loan.Reporting.Data/Models/LoanApplicationCountData.cs
@@ -2,8 +2,8 @@
 {
     public class LoanApplicationCountData
     {
-        [ColumnMapping] public short ApplicationYear { get; set; }
-        [ColumnMapping] public short ApplicationMonth { get; set; }
+        [ColumnMapping] public int ApplicationYear { get; set; }
+        [ColumnMapping] public int ApplicationMonth { get; set; }
         [ColumnMapping] public int Count { get; set; }
     }
 }
